Refactor backend architecture and improve subscription handling
- Moved AWS configuration to a separate config.js file
- Reorganized routes into a proper directory structure
- Moved trends.js to routes/trends.js for better organization
- Updated subscription metrics calculation for better performance
- Added DYNAMODB_SUBSCRIPTION_TABLE to environment variables
- Improved error handling in subscription metrics endpoint
