+++ backend/.env.example
@@ -3,4 +3,5 @@ AWS_SECRET_ACCESS_KEY=your_secret_access_key
 AWS_REGION=us-east-1
 DYNAMODB_USER_ACTIVITY_LOG_TABLE=AmazonQDevLogging
 DYNAMODB_PROMPT_LOG_TABLE=AmazonQDevPromptLog
+DYNAMODB_SUBSCRIPTION_TABLE=AmazonQDevSubscription
 PORT=3001
