+++ backend/config.js
@@ -1,0 +1,10 @@
+// Configure AWS
+const awsConfig = {
+    region: process.env.AWS_REGION || 'us-east-1',
+    accessKeyId: process.env.AWS_ACCESS_KEY_ID,
+    secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY
+};
+
+module.exports = {
+    awsConfig
+}
