+++ backend/server.js
@@ -3,8 +3,9 @@ const express = require('express');
 const cors = require('cors');
 const AWS = require('aws-sdk');
 const moment = require('moment');
-const trendsRoutes = require('./trends');
+const trendsRoutes = require('./routes/trends');
 const subscriptionsRoutes = require('./routes/subscriptions');
+const { awsConfig } = require('./config');
 
 const app = express();
 const port = process.env.PORT || 5000;
@@ -20,11 +21,7 @@ app.use(cors({
 app.use(express.json());
 
 // Configure AWS
-AWS.config.update({
-  region: process.env.AWS_REGION || 'us-east-1',
-  accessKeyId: process.env.AWS_ACCESS_KEY_ID,
-  secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY
-});
+AWS.config.update(awsConfig);
 
 const docClient = new AWS.DynamoDB.DocumentClient();
